+++ src/f10_etl/test_/test_fisc_etl_tool.py
@@ -33,8 +33,6 @@ from src.f10_etl.fisc_etl_tool import (
     get_fischour_sorted_args,
     get_fiscmont_sorted_args,
     get_fiscweek_sorted_args,
-    collect_events_dir_owner_events_sets,
-    get_owners_downhill_event_ints,
 )
 from src.f10_etl.examples.etl_env import get_test_etl_dir, env_dir_setup_cleanup
 from pandas import DataFrame, read_excel as pandas_read_excel
@@ -720,174 +718,3 @@ def test_create_fiscunit_jsons_from_prime_files_Scenario7_fisc_deal_episode(
     print(f"{expected_fiscunit.deallogs=}")
     print(f"{expected_fiscunit=}")
     assert accord56_fiscunit.deallogs == expected_fiscunit.deallogs
-
-
-def test_collect_events_dir_owner_events_sets_ReturnsObj_Scenario0_none(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fisc_events_dir = get_test_etl_dir()
-    # WHEN
-    owner_events_sets = collect_events_dir_owner_events_sets(fisc_events_dir)
-    # THEN
-    assert owner_events_sets == {}
-
-
-def test_collect_events_dir_owner_events_sets_ReturnsObj_Scenario1_DirsExist(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fisc_events_dir = get_test_etl_dir()
-    bob_str = "Bob"
-    bob_dir = create_path(fisc_events_dir, bob_str)
-    event1 = 1
-    event2 = 2
-    bob_event1_dir = create_path(bob_dir, event1)
-    bob_event2_dir = create_path(bob_dir, event2)
-    set_dir(bob_event1_dir)
-    set_dir(bob_event2_dir)
-
-    # WHEN
-    owner_events_sets = collect_events_dir_owner_events_sets(fisc_events_dir)
-
-    # THEN
-    assert owner_events_sets == {bob_str: {event1, event2}}
-
-
-def test_collect_events_dir_owner_events_sets_ReturnsObj_Scenario2_DirsExist(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fisc_events_dir = get_test_etl_dir()
-    bob_str = "Bob"
-    sue_str = "Sue"
-    bob_dir = create_path(fisc_events_dir, bob_str)
-    sue_dir = create_path(fisc_events_dir, sue_str)
-    event1 = 1
-    event2 = 2
-    event7 = 7
-    bob_event1_dir = create_path(bob_dir, event1)
-    bob_event2_dir = create_path(bob_dir, event2)
-    sue_event2_dir = create_path(sue_dir, event2)
-    sue_event7_dir = create_path(sue_dir, event7)
-    set_dir(bob_event1_dir)
-    set_dir(bob_event2_dir)
-    set_dir(sue_event2_dir)
-    set_dir(sue_event7_dir)
-
-    # WHEN
-    owner_events_sets = collect_events_dir_owner_events_sets(fisc_events_dir)
-
-    # THEN
-    assert owner_events_sets == {bob_str: {event1, event2}, sue_str: {event2, event7}}
-
-
-def test_get_owners_downhill_event_ints_ReturnsObj_Scenario0_Empty():
-    # ESTABLISH
-    bob_str = "Bob"
-    sue_str = "Sue"
-    event2 = 2
-    owner_events_sets = {}
-    downhill_event_int = event2
-    downhill_owners = {bob_str, sue_str}
-
-    # WHEN
-    owners_downhill_event_ints = get_owners_downhill_event_ints(
-        owner_events_sets, downhill_owners, downhill_event_int
-    )
-
-    # THEN
-    assert owners_downhill_event_ints == {}
-
-
-def test_get_owners_downhill_event_ints_ReturnsObj_Scenario1_simple():
-    # ESTABLISH
-    bob_str = "Bob"
-    sue_str = "Sue"
-    event1 = 1
-    event2 = 2
-    event7 = 7
-    owner_events_sets = {bob_str: {event1, event2}, sue_str: {event2, event7}}
-    downhill_event_int = event2
-    downhill_owners = {bob_str, sue_str}
-
-    # WHEN
-    owners_downhill_event_ints = get_owners_downhill_event_ints(
-        owner_events_sets, downhill_owners, downhill_event_int
-    )
-
-    # THEN
-    assert owners_downhill_event_ints == {bob_str: event2, sue_str: event2}
-
-
-def test_get_owners_downhill_event_ints_ReturnsObj_Scenario2Empty_downhill_event_int():
-    # ESTABLISH
-    bob_str = "Bob"
-    sue_str = "Sue"
-    yao_str = "Yao"
-    event1 = 1
-    event2 = 2
-    event7 = 7
-    owner_events_sets = {
-        bob_str: {event1, event2},
-        sue_str: {event2, event7},
-        yao_str: {event1, event2, event7},
-    }
-    downhill_owners = {bob_str, sue_str}
-
-    # WHEN
-    owners_downhill_event_ints = get_owners_downhill_event_ints(
-        owner_events_sets, downhill_owners
-    )
-
-    # THEN
-    assert owners_downhill_event_ints == {bob_str: event2, sue_str: event7}
-
-
-def test_get_owners_downhill_event_ints_ReturnsObj_Scenario3Empty_downhill_owners():
-    # ESTABLISH
-    bob_str = "Bob"
-    sue_str = "Sue"
-    yao_str = "Yao"
-    event1 = 1
-    event2 = 2
-    event7 = 7
-    owner_events_sets = {
-        bob_str: {event1, event2},
-        sue_str: {event2, event7},
-        yao_str: {event1, event2, event7},
-    }
-
-    # WHEN
-    owners_downhill_event_ints = get_owners_downhill_event_ints(owner_events_sets)
-
-    # THEN
-    assert owners_downhill_event_ints == {
-        bob_str: event2,
-        sue_str: event7,
-        yao_str: event7,
-    }
-
-
-def test_get_owners_downhill_event_ints_ReturnsObj_Scenario4Empty_downhill_owners_Withdownhill_event_int():
-    # ESTABLISH
-    bob_str = "Bob"
-    sue_str = "Sue"
-    yao_str = "Yao"
-    event1 = 1
-    event2 = 2
-    event7 = 7
-    owner_events_sets = {
-        bob_str: {event1, event2},
-        sue_str: {event2, event7},
-        yao_str: {event7},
-    }
-    downhill_event_int = 2
-
-    # WHEN
-    owners_downhill_event_ints = get_owners_downhill_event_ints(
-        owner_events_sets, ref_event_int=downhill_event_int
-    )
-
-    # THEN
-    assert owners_downhill_event_ints == {bob_str: event2, sue_str: event2}
